+++ src/nti/site/runner.py
@@ -18,7 +18,7 @@ from zope.component.hooks import site as current_site
 
 from ZODB.interfaces import IDatabase
 
-from nti.transactions.transactions import TransactionLoop
+from nti.transactions.loop import TransactionLoop
 
 from nti.site.interfaces import SiteNotInstalledError